Multi-pdf-chatbot/requirements.txt
langchain==0.0.184
PyPDF2==3.0.1
python-dotenv==1.0.0
streamlit==1.18.1
openai==0.27.6
faiss-cpu==1.7.3       # ✅ Downgraded to avoid SWIG requirement
altair==4.2.2           # ✅ Fully specified version
tiktoken==0.3.3         # ✅ Downgraded to avoid needing Rust
huggingface-hub==0.14.1


# uncomment to use instructor embeddings
# InstructorEmbedding==1.0.1
# sentence-transformers==2.2.2